Do section filtering always at runtime
Remove section filtering from `section-loader` that is performed at build time.

Create a new module called `sections-filter`. It contains code that will run in both Node.js
and browser. `server/bundler/utils` has Node-specific code that depends on modules like `crypto`
or `fs`.

Use the `isSectionEnabled` function to filter the sections list at runtime, both on server and client.
It will filter according to the `config` that is active at runtime.
