+++ client/sections-middleware.js
@@ -14,6 +14,7 @@ import * as LoadingError from 'layout/error';
 import * as controller from './controller/index.web';
 import { pathToRegExp } from './utils';
 import { receiveSections, load } from './sections-helper';
+import isSectionEnabled from './section-filter';
 import { addReducerToStore } from 'state/add-reducer';
 
 import sections from './sections';
@@ -106,6 +107,10 @@ function createPageDefinition( path, sectionDefinition ) {
 
 export const setupRoutes = () => {
 	for ( const section of sections ) {
+		if ( ! isSectionEnabled( section ) ) {
+			continue;
+		}
+
 		for ( const path of section.paths ) {
 			createPageDefinition( path, section );
 		}
